+++ opencti-platform/opencti-front/src/private/components/threats/threat_actors_individual/AddPersonasThreatActorIndividualLines.tsx
@@ -1,15 +1,15 @@
-import React, { FunctionComponent, useEffect, useState } from 'react';
-import { graphql, useRefetchableFragment } from 'react-relay';
+import React, { FunctionComponent, useState } from 'react';
+import { graphql, useFragment } from 'react-relay';
 import Loader from 'src/components/Loader';
 import { List, ListItemButton, ListItemIcon, ListItemText, useTheme } from '@mui/material';
 import ItemIcon from 'src/components/ItemIcon';
 import { CheckCircle } from '@mui/icons-material';
 import useApiMutation from 'src/utils/hooks/useApiMutation';
 import { defaultCommitMutation } from 'src/relay/environment';
-import { RecordSourceSelectorProxy } from 'relay-runtime';
 import { scoRelationshipAdd, scoRelationshipDelete } from '@components/threats/threat_actors_individual/AddIndividualsThreatActorIndividualLines';
 import { ThreatActorIndividualDetails_ThreatActorIndividual$data } from './__generated__/ThreatActorIndividualDetails_ThreatActorIndividual.graphql';
 import { AddPersonasThreatActorIndividualLines_data$key } from './__generated__/AddPersonasThreatActorIndividualLines_data.graphql';
+import { updateDelete } from '../../../../utils/store';
 
 export const addPersonasThreatActorIndividualLinesQuery = graphql`
   query AddPersonasThreatActorIndividualLinesQuery(
@@ -94,14 +94,10 @@ AddPersonasThreatActorIndividualLinesProps
   threatActorIndividual,
   fragmentKey,
 }) => {
-  const [data, refetch] = useRefetchableFragment(
+  const data = useFragment(
     AddPersonasThreatActorIndividualLinesFragment,
     fragmentKey,
   );
-
-  useEffect(() => {
-    refetch({});
-  }, [data]);
 
   const [commitRelationAdd] = useApiMutation(scoRelationshipAdd);
   const [commitRelationDelete] = useApiMutation(scoRelationshipDelete);
@@ -112,15 +108,6 @@ AddPersonasThreatActorIndividualLinesProps
     ?? []).map(({ node }) => node.to?.id ?? '');
 
   const [currentTargets, setCurrentTargets] = useState<string[]>(initialTargets);
-
-  const updateDelete = (store: RecordSourceSelectorProxy, path: string, rootId: string, deleteId: string) => {
-    const node = store.get(rootId);
-    const records = node?.getLinkedRecord(path);
-    const edges = records?.getLinkedRecords('edges');
-    if (!records || !edges) { return; }
-    const newEdges = edges.filter((n) => n.getLinkedRecord('node')?.getValue('toId') !== deleteId);
-    records.setLinkedRecords(newEdges, 'edges');
-  };
 
   const handleToggle = (toId: string) => {
     const isSelected = currentTargets.includes(toId);
