+++ opencti-platform/opencti-front/src/private/components/threats/threat_actors_individual/AddThreatActorIndividualDemographic.tsx
@@ -4,42 +4,42 @@ import IconButton from '@mui/material/IconButton';
 import { Add } from '@mui/icons-material';
 import Drawer from '@components/common/drawer/Drawer';
 import CountryCreation from '@components/locations/countries/CountryCreation';
-import { useLazyLoadQuery } from 'react-relay';
 import AddThreatActorIndividualDemographicLines, {
   addIndividualsThreatActorIndividualLinesQuery,
 } from '@components/threats/threat_actors_individual/AddThreatActorIndividualDemographicLines';
 import { AddThreatActorIndividualDemographicLinesQuery } from '@components/threats/threat_actors_individual/__generated__/AddThreatActorIndividualDemographicLinesQuery.graphql';
+import { PreloadedQuery, usePreloadedQuery } from 'react-relay';
 import { useFormatter } from '../../../../components/i18n';
 import SearchInput from '../../../../components/SearchInput';
+import useQueryLoading from '../../../../utils/hooks/useQueryLoading';
+import Loader, { LoaderVariant } from '../../../../components/Loader';
 
 interface AddThreatActorIndividualDemographicProps {
   threatActorIndividual: ThreatActorIndividual_ThreatActorIndividual$data,
   relType: string,
   title:string,
+  queryRef: PreloadedQuery<AddThreatActorIndividualDemographicLinesQuery>;
 }
 
-const AddThreatActorIndividualDemographic: FunctionComponent<
+const AddThreatActorIndividualDemographicComponent: FunctionComponent<
 AddThreatActorIndividualDemographicProps
 > = ({
   threatActorIndividual,
   relType,
   title,
+  queryRef,
 }) => {
   const { t_i18n } = useFormatter();
   const [open, setOpen] = useState<boolean>(false);
   const [search, setSearch] = useState<string>('');
-  const paginationOptions = { search };
 
   const handleOpen = () => setOpen(true);
   const handleClose = () => setOpen(false);
   const handleSearch = (term: string) => setSearch(term);
 
-  const data = useLazyLoadQuery<AddThreatActorIndividualDemographicLinesQuery>(
+  const data = usePreloadedQuery<AddThreatActorIndividualDemographicLinesQuery>(
     addIndividualsThreatActorIndividualLinesQuery,
-    {
-      ...paginationOptions,
-      count: 50,
-    },
+    queryRef,
   );
 
   return (<div>
@@ -69,7 +69,6 @@ AddThreatActorIndividualDemographicProps
             variant='inDrawer'
             onSubmit={handleSearch}
           />
-          <div style={{ height: 5 }} />
           <CountryCreation
             paginationOptions={{
               search,
@@ -86,6 +85,21 @@ AddThreatActorIndividualDemographicProps
       />
     </Drawer>
   </div>);
+};
+
+const AddThreatActorIndividualDemographic: FunctionComponent<
+Omit<AddThreatActorIndividualDemographicProps, 'queryRef'>
+> = (props) => {
+  const queryRef = useQueryLoading<AddThreatActorIndividualDemographicLinesQuery>(addIndividualsThreatActorIndividualLinesQuery, {
+    count: 50,
+  });
+  return queryRef ? (
+    <React.Suspense fallback={<Loader variant={LoaderVariant.inElement} />}>
+      <AddThreatActorIndividualDemographicComponent {...props} queryRef={queryRef} />
+    </React.Suspense>
+  ) : (
+    <Loader variant={LoaderVariant.inElement} />
+  );
 };
 
 export default AddThreatActorIndividualDemographic;
